Update .env.sample and README.md
Update the DATABASE_URL in .env.sample

The DATABASE_URL in .env.sample was pointing to the wrong database. This commit updates it to point to the correct database.

Also, update the link in README.md

The link in README.md was pointing to the wrong location for setting up the Farcaster PosgreSQL indexer. This commit updates the link to point to the correct location.